requirements.txt
lightning-flash==0.7.5
ray[tune]==1.12.1
streamlit==1.9.0
